Move game AI service under storyteller module
